Move schemas related browse code to separate folder
