+++ frontend/src/metabase/browse/index.tsx
@@ -2,4 +2,4 @@ export { BrowseMetrics } from "./metrics";
 export { BrowseModels } from "./models";
 export { BrowseDatabases } from "./components/BrowseDatabases";
 export { BrowseTables } from "./components/BrowseTables";
-export { default as BrowseSchemas } from "./components/BrowseSchemas";
+export { BrowseSchemas } from "./schemas";
