+++ frontend/src/metabase/browse/schemas/index.tsx
@@ -1,0 +1,1 @@
+export { BrowseSchemas } from "./BrowseSchemas";
